- Arreglado el messagepump de nuevo ... muy a lo hack pero es que ogre y flash se llevan como el culo xD - Puesto el heavy del tf2 como enemigo, como personaje principal de momento no hasta que arregle las texturas - Mensaje para antonio: al meter el heavy he jodido parte de las texturas del mapa porque se llaman igual, tenemos que arreglarlo!

+++ Projects/Hikari/source/ViewportOverlay.cpp
@@ -20,7 +20,9 @@
 */
 
 #include "impl/ViewportOverlay.h"
-
+#include <OgreOverlay.h>
+#include <OgreViewport.h>
+#include <OgreOverlayManager.h>
 using namespace Ogre;
 using namespace Hikari;
 using namespace Hikari::Impl;
@@ -41,19 +43,19 @@ ViewportOverlay::ViewportOverlay(const Ogre::String& name, Ogre::Viewport* viewp
 	overlay->setZOrder(zOrder);
 	resetPosition();
 
-	viewport->getTarget()->addListener(this);
+	//viewport->getTarget()->addListener(this);
 }
 
 ViewportOverlay::~ViewportOverlay()
 {
-	viewport->getTarget()->removeListener(this);
+	//viewport->getTarget()->removeListener(this);
 
-	//if(overlay)
-	//{
+	if(overlay)
+	{
 		overlay->remove2D(panel);
 		OverlayManager::getSingletonPtr()->destroyOverlayElement(panel);
 		OverlayManager::getSingletonPtr()->destroy(overlay);
-	//}
+	}
 }
 
 void ViewportOverlay::move(int deltaX, int deltaY)
@@ -130,11 +132,13 @@ void ViewportOverlay::resize(int width, int height)
 void ViewportOverlay::hide()
 {
 	isVisible = false;
+	overlay->hide();
 }
 
 void ViewportOverlay::show()
 {
 	isVisible = true;
+	overlay->show();
 }
 
 int ViewportOverlay::getRelativeX(int absX)
